requirements.txt
streamlit
pinecone-client==2.2.4
langchain-community==0.0.29
openai>=1.12.0
python-dotenv>=1.0.0